Change Activty to be in System.Diagnostics rather than System.Diagnostics.Activity
There are ambiguities that arise if you name a class the same name as the namespace that contains it.  We should not do this.
I don't think we need a namespace for Activies (there seems to be only one class), so the obvious answer is to simply move
Activity to the System.Diagnostic namespace.